fix getblock integer overflow for buffers over 32GB in size

+++ MurmurHash3.cpp
@@ -57,12 +57,12 @@ inline uint64_t rotl64 ( uint64_t x, int8_t r )
 // Block read - if your platform needs to do endian-swapping or can only
 // handle aligned reads, do the conversion here
 
-FORCE_INLINE uint32_t getblock ( const uint32_t * p, int i )
+FORCE_INLINE uint32_t getblock ( const uint32_t * p, Py_ssize_t i)
 {
   return p[i];
 }
 
-FORCE_INLINE uint64_t getblock ( const uint64_t * p, int i )
+FORCE_INLINE uint64_t getblock ( const uint64_t * p, Py_ssize_t i )
 {
   return p[i];
 }
